+++ lib/src/core/models/chat/sub_models/choices/sub_models/message.dart
@@ -1,5 +1,6 @@
 // ignore_for_file: public_member_api_docs, sort_constructors_first
-import '../../../../image/enum.dart';
+import '../../../../../enum.dart';
+import '../../../etc/message_adapter.dart';
 import 'sub_models/content.dart';
 import 'sub_models/tool_call.dart';
 export 'sub_models/content.dart';
@@ -12,15 +13,20 @@ final class OpenAIChatCompletionChoiceMessageModel {
   /// The [role] of the message.
   final OpenAIChatMessageRole role;
 
-  //! TODO: add the possibility to include images in the message content, see docs and do a non-breaking change.
-
   /// The [content] of the message.
   final List<OpenAIChatCompletionChoiceMessageContentItemModel>? content;
 
   /// The function that the model is requesting to call.
   final List<OpenAIResponseToolCall>? toolCalls;
 
-  bool get hasToolCalls => toolCalls != null;
+  /// The message participent name.
+  final String? name;
+
+  /// Weither the message have tool calls.
+  bool get haveToolCalls => toolCalls != null;
+
+  /// Weither the message have content.
+  bool get haveContent => content != null && content!.isNotEmpty;
 
   @override
   int get hashCode {
@@ -32,6 +38,7 @@ final class OpenAIChatCompletionChoiceMessageModel {
     required this.role,
     required this.content,
     this.toolCalls,
+    this.name,
   });
 
   /// This is used  to convert a [Map<String, dynamic>] object to a [OpenAIChatCompletionChoiceMessageModel] object.
@@ -39,10 +46,13 @@ final class OpenAIChatCompletionChoiceMessageModel {
     Map<String, dynamic> json,
   ) {
     return OpenAIChatCompletionChoiceMessageModel(
+      name: json['name'],
       role: OpenAIChatMessageRole.values
           .firstWhere((role) => role.name == json['role']),
       content: json['content'] != null
-          ? dynamicContentFromField(json['content'])
+          ? OpenAIMessageDynamicContentFromFieldAdapter.dynamicContentFromField(
+              json['content'],
+            )
           : null,
       toolCalls: json['tool_calls'] != null
           ? (json['tool_calls'] as List)
@@ -59,6 +69,7 @@ final class OpenAIChatCompletionChoiceMessageModel {
       "content": content?.map((contentItem) => contentItem.toMap()).toList(),
       if (toolCalls != null && role == OpenAIChatMessageRole.assistant)
         "tool_calls": toolCalls!.map((toolCall) => toolCall.toMap()).toList(),
+      if (name != null) "name": name,
     };
   }
 
@@ -86,8 +97,9 @@ final class OpenAIChatCompletionChoiceMessageModel {
         other.toolCalls == toolCalls;
   }
 
-//! TODO: make method for all other kind of resending messages.
-
+  /// Converts a response function message to a request function message, so that it can be used in the next request.
+  ///
+  /// You should pass the response function message's [toolCallId] to this method, since it is required when requesting it.
   RequestFunctionMessage asRequestFunctionMessage({
     required String toolCallId,
   }) {
@@ -97,39 +109,17 @@ final class OpenAIChatCompletionChoiceMessageModel {
       toolCallId: toolCallId,
     );
   }
-
-  static List<OpenAIChatCompletionChoiceMessageContentItemModel>
-      dynamicContentFromField(
-    fieldData,
-  ) {
-    if (fieldData is String) {
-      return [
-        OpenAIChatCompletionChoiceMessageContentItemModel.text(fieldData),
-      ];
-    } else if (fieldData is List) {
-      return (fieldData).map(
-        (item) {
-          if (item is! Map) {
-            throw Exception('Invalid content item');
-          } else {
-            final asMap = item as Map<String, dynamic>;
-
-            return OpenAIChatCompletionChoiceMessageContentItemModel.fromMap(
-              asMap,
-            );
-          }
-        },
-      ).toList();
-    } else {
-      throw Exception('Invalid content');
-    }
-  }
 }
 
+/// {@template openai_chat_completion_function_choice_message_model}
+/// This represents the message of the [RequestFunctionMessage] model of the OpenAI API, which is used  while using the [OpenAIChat] methods, precisely to send a response function message as a request function message for next requests.
+/// {@endtemplate}
 base class RequestFunctionMessage
     extends OpenAIChatCompletionChoiceMessageModel {
+  /// The [toolCallId] of the message.
   final String toolCallId;
 
+  /// {@macro openai_chat_completion_function_choice_message_model}
   RequestFunctionMessage({
     required super.role,
     required super.content,
@@ -140,8 +130,10 @@ base class RequestFunctionMessage
   Map<String, dynamic> toMap() {
     return {
       "role": role.name,
-      "content": content,
+      "content": content?.map((toolCall) => toolCall.toMap()).toList(),
       "tool_call_id": toolCallId,
     };
   }
+
+  //! Does this needs fromMap method?
 }