+++ lib/src/instance/chat/chat.dart
@@ -82,6 +82,8 @@ interface class OpenAIChat implements OpenAIChatBase {
     String? user,
     Map<String, String>? responseFormat,
     int? seed,
+    bool? logprobs,
+    int? topLogprobs,
     http.Client? client,
   }) async {
     return await OpenAINetworkingClient.post(
@@ -91,7 +93,7 @@ interface class OpenAIChat implements OpenAIChatBase {
         "messages": messages.map((message) => message.toMap()).toList(),
         if (tools != null)
           "tools": tools.map((tool) => tool.toMap()).toList(growable: false),
-        if (toolChoice != null) "tool_choice": toolChoice.value,
+        if (toolChoice != null) "tool_choice": toolChoice,
         if (temperature != null) "temperature": temperature,
         if (topP != null) "top_p": topP,
         if (n != null) "n": n,
@@ -103,6 +105,8 @@ interface class OpenAIChat implements OpenAIChatBase {
         if (user != null) "user": user,
         if (seed != null) "seed": seed,
         if (responseFormat != null) "response_format": responseFormat,
+        if (logprobs != null) "logprobs": logprobs,
+        if (topLogprobs != null) "top_logprobs": topLogprobs,
       },
       onSuccess: (Map<String, dynamic> response) {
         return OpenAIChatCompletionModel.fromMap(response);
@@ -185,7 +189,7 @@ interface class OpenAIChat implements OpenAIChatBase {
         "messages": messages.map((message) => message.toMap()).toList(),
         if (tools != null)
           "tools": tools.map((tool) => tool.toMap()).toList(growable: false),
-        if (toolChoice != null) "tool_choice": toolChoice.value,
+        if (toolChoice != null) "tool_choice": toolChoice,
         if (temperature != null) "temperature": temperature,
         if (topP != null) "top_p": topP,
         if (n != null) "n": n,
@@ -232,7 +236,7 @@ interface class OpenAIChat implements OpenAIChatBase {
         "messages": messages.map((message) => message.toMap()).toList(),
         if (tools != null)
           "tools": tools.map((tool) => tool.toMap()).toList(growable: false),
-        if (toolChoice != null) "tool_choice": toolChoice.value,
+        if (toolChoice != null) "tool_choice": toolChoice,
         if (temperature != null) "temperature": temperature,
         if (topP != null) "top_p": topP,
         if (n != null) "n": n,