+++ test/openai_test.dart
@@ -1,8 +1,10 @@
 // ignore_for_file: avoid-passing-async-when-sync-expected
 
+import 'dart:async';
 import 'dart:convert';
 import 'dart:io';
 import 'dart:developer' as dev;
+import 'package:collection/collection.dart';
 import 'package:dart_openai/dart_openai.dart';
 import 'package:dart_openai/src/core/builder/headers.dart';
 import 'package:dart_openai/src/core/constants/config.dart';
@@ -16,15 +18,9 @@ void main() async {
   final exampleImageFile = await getFileFromUrl(
     "https://upload.wikimedia.org/wikipedia/commons/7/7e/Dart-logo.png",
   );
-  final audioExampleFile = await getFileFromUrl(
-    "https://www.cbvoiceovers.com/wp-content/uploads/2017/05/Commercial-showreel.mp3",
-    fileExtension: "mp3",
-  );
 
   final imageFileExample = await exampleImageFile;
   final maskFileExample = await exampleImageFile;
-
-  String? modelExampleId;
 
   String? fileIdFromFilesApi;
 
@@ -106,6 +102,8 @@ void main() async {
   });
 
   group('models', () {
+    String? modelExampleId;
+
     test(
       'list models',
       () async {
@@ -118,16 +116,21 @@ void main() async {
           expect(models.first.id, isNotNull);
           expect(models.first.id, isA<String>());
 
-          if (models.first.permission != null) {
-            expect(
-              models.first.permission,
-              isA<List<OpenAIModelModelPermission>>(),
-            );
+          final someModel = models.first;
+
+          if (someModel.permission != null &&
+              someModel.permission!.isNotEmpty) {
+            final permission = someModel.permission!.first;
+
+            expect(permission, isA<OpenAIModelModelPermission>());
+
+            expect(permission.id, isNotNull);
           }
 
-          // trying to get the "text-davinci-003" model id.
           modelExampleId = models
-              .firstWhere((element) => element.id.contains("davinci-003"))
+              .firstWhereOrNull(
+                (element) => element.id.contains("gpt-3"),
+              )!
               .id;
         }
       },
@@ -138,6 +141,7 @@ void main() async {
         modelExampleId != null,
         "please set a model id that is not null, or let the previous test run first to get a model id example",
       );
+
       final model = await OpenAI.instance.model.retrieve(modelExampleId!);
 
       expect(model, isA<OpenAIModelModel>());
@@ -149,15 +153,23 @@ void main() async {
           model.permission,
           isA<List<OpenAIModelModelPermission>>(),
         );
+
+        if (model.permission!.isNotEmpty) {
+          final permission = model.permission!.first;
+
+          expect(permission, isA<OpenAIModelModelPermission>());
+
+          expect(permission.id, isNotNull);
+        }
       }
     });
   });
+
   group('completions', () {
     test('create', () async {
       final OpenAICompletionModel completion =
           await OpenAI.instance.completion.create(
-        // in case the previous test didn't run, we will use a default model id.
-        model: modelExampleId ?? "text-davinci-003",
+        model: "davinci-002",
         prompt: "Dart tests are made to ensure that a function w",
         maxTokens: 5,
         temperature: 0.9,
@@ -174,11 +186,10 @@ void main() async {
       expect(completion.choices.first.text, isA<String?>());
     });
 
-    test('create with a stream', () {
-      final Stream<OpenAIStreamCompletionModel> completion =
+    test('create with a stream', () async {
+      final Stream<OpenAIStreamCompletionModel> completionStream =
           OpenAI.instance.completion.createStream(
-        // in case the previous test didn't run, we will use a default model id.
-        model: modelExampleId ?? "text-davinci-003",
+        model: "davinci-002",
         prompt: "Dart tests are made to ensure that a function w",
         maxTokens: 5,
         temperature: 0.9,
@@ -188,7 +199,22 @@ void main() async {
         bestOf: 1,
         n: 1,
       );
-      expect(completion, isA<Stream<OpenAIStreamCompletionModel>>());
+
+      final completer = Completer<bool>();
+      expect(completionStream, isA<Stream<OpenAIStreamCompletionModel>>());
+
+      completionStream.listen(
+        (event) {
+          var val = event.choices.first.text;
+
+          expect(val, isA<String>());
+        },
+        onDone: () {
+          completer.complete(true);
+        },
+      );
+
+      await completer.future;
     });
   });
 
@@ -219,7 +245,16 @@ void main() async {
         chatCompletion.choices.first.message,
         isA<OpenAIChatCompletionChoiceMessageModel>(),
       );
-      expect(chatCompletion.choices.first.message.content, isA<String>());
+      expect(
+        chatCompletion.choices.first.message.content,
+        isA<List<OpenAIChatCompletionChoiceMessageContentItemModel>>(),
+      );
+      expect(
+        chatCompletion.choices.first.message.content!
+            .map((e) => e.text)
+            .isEmpty,
+        isFalse,
+      );
     });
 
     test(
@@ -233,7 +268,7 @@ void main() async {
 
         final OpenAIChatCompletionModel chatCompletion =
             await OpenAI.instance.chat.create(
-          model: "gpt-3.5-turbo-0613",
+          model: "gpt-3.5-turbo",
           messages: [
             OpenAIChatCompletionChoiceMessageModel(
               content: [
@@ -293,7 +328,7 @@ void main() async {
       },
     );
 
-    test('create with a stream', () {
+    test('create with a stream', () async {
       final chatStream = OpenAI.instance.chat.createStream(
         model: "gpt-3.5-turbo",
         messages: [
@@ -307,27 +342,26 @@ void main() async {
           ),
         ],
       );
+      final completer = Completer<bool>();
+
       expect(chatStream, isA<Stream<OpenAIStreamChatCompletionModel>>());
-      chatStream.listen((streamEvent) {
-        expect(streamEvent, isA<OpenAIStreamChatCompletionModel>());
-        expect(streamEvent.choices.first.delta.content, isA<String?>());
-      });
-    });
-  });
-  group('edits', () {
-    // ! temporary disabled, because the API have on this and throws an unexpected error from OpenAI end.
-    test('create', () async {
-      final OpenAIEditModel edit = await OpenAI.instance.edit.create(
-        model: "text-davinci-edit-001",
-        instruction: "remove the word 'made' from the text",
-        input: "I made something, idk man",
+      chatStream.listen(
+        (streamEvent) {
+          expect(streamEvent, isA<OpenAIStreamChatCompletionModel>());
+          expect(streamEvent.choices.first.delta.content,
+              isA<List<OpenAIChatCompletionChoiceMessageContentItemModel>?>());
+          expect(streamEvent.choices.first.delta.content?.first?.text,
+              isA<String?>());
+        },
+        onDone: () {
+          completer.complete(true);
+        },
       );
-      expect(edit, isA<OpenAIEditModel>());
-      expect(edit.choices.first, isA<OpenAIEditModelChoice>());
-      expect(edit.choices.first.text, isNotNull);
-      expect(edit.choices.first.text, isA<String>());
+
+      await completer.future;
     });
   });
+
   group('images', () {
     test('create', () async {
       final OpenAIImageModel image = await OpenAI.instance.image.create(
@@ -371,16 +405,43 @@ void main() async {
 
   group('audio', () {
     test("create transcription", () async {
+      final audioExampleFile = await getFileFromUrl(
+        "https://www.cbvoiceovers.com/wp-content/uploads/2017/05/Commercial-showreel.mp3",
+        fileExtension: "mp3",
+      );
+
       final transcription = await OpenAI.instance.audio.createTranscription(
         file: audioExampleFile,
         model: "whisper-1",
         responseFormat: OpenAIAudioResponseFormat.json,
       );
-
       expect(transcription, isA<OpenAIAudioModel>());
       expect(transcription.text, isA<String>());
     });
+
+    test("create transcription with timestamp granularity", () async {
+      final audioExampleFile = await getFileFromUrl(
+        "https://www.cbvoiceovers.com/wp-content/uploads/2017/05/Commercial-showreel.mp3",
+        fileExtension: "mp3",
+      );
+
+      final transcription = await OpenAI.instance.audio.createTranscription(
+        file: audioExampleFile,
+        model: "whisper-1",
+        responseFormat: OpenAIAudioResponseFormat.verbose_json,
+        timestamp_granularities: [OpenAIAudioTimestampGranularity.word],
+      );
+
+      expect(transcription, isA<OpenAIAudioModel>());
+      expect(transcription.text, isA<String>());
+      expect(transcription.words, isA<List>());
+    });
     test("create translation", () async {
+      final audioExampleFile = await getFileFromUrl(
+        "https://www.cbvoiceovers.com/wp-content/uploads/2017/05/Commercial-showreel.mp3",
+        fileExtension: "mp3",
+      );
+
       final translation = await OpenAI.instance.audio.createTranslation(
         file: audioExampleFile,
         model: "whisper-1",
@@ -529,9 +590,12 @@ void main() async {
 File jsonLFileExample() {
   final file = File("example.jsonl");
   file.writeAsStringSync(
-    jsonEncode(
-        {"prompt": "<prompt text>", "completion": "<ideal generated text>"}),
+    jsonEncode({
+      "prompt": "<prompt text>",
+      "completion": "<ideal generated text>",
+    }),
   );
+
   return file;
 }
 
@@ -543,5 +607,6 @@ Future<File> getFileFromUrl(
   final uniqueImageName = DateTime.now().microsecondsSinceEpoch;
   final file = File("$uniqueImageName.$fileExtension");
   await file.writeAsBytes(response.bodyBytes);
+
   return file;
 }